requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
ec==0.2.5
exceptiongroup==1.1.0
fastapi==0.89.1
h11==0.14.0
idna==3.4
imgur-uploader==0.3.0
imgurpython==1.1.7
outcome==1.2.0
packaging==23.0
pydantic==1.10.4
PyJWT==2.6.0
pyotp==2.8.0
pyperclip==1.8.2
PySocks==1.7.1
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
starlette==0.22.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
twilio==7.16.2
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
webdriver-manager==3.8.5
wsproto==1.2.0
